requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-osc>=1.8.3
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.26.0
mss>=9.0.1
requests>=2.31.0
openai>=1.12.0
Pillow>=10.2.0
scipy>=1.12.0
soundfile>=0.12.1
python-multipart>=0.0.9
websockets>=12.0
transformers>=4.38.0
# torch с CUDA — устанавливается отдельно (нужен конкретный индекс под GPU):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
